Working prototype of major functionality
This borrows heavily from css-tricks form validation, which was the
most helpful resource I found in fully grasping the Constraint Validation
API. Created after a bit of a hiatus due to graduate school, it is
an inelegant but working solution. Certainly it abandons some of the
previous work with containers, but effectively prints and clears
relevant messages, using dependent logic based on selections/entries
of other input fields (e.g., Zip is dependent on Country, confirmation
on password). My concern specifically about setting the pattern in the
way that I did for confirmation is that this might expose a user's
password unnecessarily, and likely should be handled more internally.
However, given my lack of familiariaty with Constraint Validation API,
I've decided to hold off on that until further reading.